LPS-57070 Rename to search-facets-api
