+++ modules/apps/search/search-facets-api/src/com/liferay/search/facets/api/util/SearchFacetTracker.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.search.api.util;
+package com.liferay.search.facets.api.util;
 
-import com.liferay.search.api.facet.SearchFacet;
+import com.liferay.search.facets.api.facet.SearchFacet;
 
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
