+++ modules/apps/search/search-web/build.gradle
@@ -1,6 +1,6 @@
 dependencies {
 	compile group: "org.osgi", name: "org.osgi.compendium", version: "5.0.0"
-	compile project(":apps:search:search-api")
+	compile project(":apps:search:search-facets-api")
 }
 
 liferay {
