+++ modules/apps/search/search-web/src/com/liferay/search/web/facet/AssetEntriesSearchFacet.java
@@ -27,8 +27,8 @@ import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portal.util.WebKeys;
 import com.liferay.portlet.asset.AssetRendererFactoryRegistryUtil;
 import com.liferay.portlet.asset.model.AssetRendererFactory;
-import com.liferay.search.api.facet.BaseJSPSearchFacet;
-import com.liferay.search.api.facet.SearchFacet;
+import com.liferay.search.facets.api.facet.BaseJSPSearchFacet;
+import com.liferay.search.facets.api.facet.SearchFacet;
 
 import java.util.ArrayList;
 import java.util.List;
